1.0

+++ backend/routes/api.php
@@ -1,38 +1,34 @@
 <?php
-
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider and all of them will
-| be assigned to the "api" middleware group. Make something great!
-|
-*/
-
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
+Route::middleware('auth:api')->get('/user', function (Request $request) {
+    return response()->json($request->user());
 });
 
 Route::controller(\App\Http\Controllers\PeticioneController::class)->group(function () {
     Route::get('peticiones', 'index');
     Route::get('mispeticiones', 'listmine');
     Route::get('peticiones/{id}', 'show');
-    Route::delete('peticiones/{id}', 'delete');
-    Route::put('peticiones/firmar/{id}', 'firmar');
-    Route::put('peticiones/{id}', 'update');
-    Route::put('peticiones/estado/{id}', 'cambiarEstado');
-    Route::post('peticiones', 'store');
+    Route::delete('peticiones/{id}', 'delete')->middleware('auth:api'); // Proteger con JWT
+    Route::put('peticiones/firmar/{id}', 'firmar')->middleware('auth:api'); // Proteger con JWT
+    Route::put('peticiones/{id}', 'update')->middleware('auth:api'); // Proteger con JWT
+    Route::put('peticiones/estado/{id}', 'cambiarEstado')->middleware('auth:api'); // Proteger con JWT
+    Route::post('peticiones', 'store')->middleware('auth:api'); // Proteger con JWT
+    Route::get('misfirmas', 'listSigned')->middleware('auth:api'); // Proteger con JWT
 });
 
 Route::controller(\App\Http\Controllers\AuthController::class)->group(function () {
     Route::post('login', 'login');
     Route::post('register', 'register');
-    Route::post('logout', 'logout');
-    Route::post('refresh', 'refresh');
-    Route::get('me', 'me');
-    });
+    Route::post('logout', 'logout')->middleware('auth:api'); // Proteger con JWT
+    Route::post('refresh', 'refresh')->middleware('auth:api'); // Proteger con JWT
+    Route::get('me', 'me')->middleware('auth:api'); // Proteger con JWT
+});
+
+
+Route::post('test-update', function (Request $request) {
+    return response()->json([
+        'received_data' => $request->all(),
+        'received_files' => $request->file('files'),
+    ], 200);
+});
